runtime/tests: Move computed.collect() test to corresponding test file
